+++ scripts/background.js
@@ -77,7 +77,7 @@ const fetchReviewedKanjiID = async (apiToken, page) => {
 // transform the kanji IDs into kanji characters
 const setupLearnedKanji = async (apiToken, page, kanji) => {
 	const IDs = await fetchReviewedKanjiID(apiToken, page);
-	const learnedKanji = IDs.map(id => kanji["wkhighlight_allkanji"][id].slug);
+	const learnedKanji = IDs.map(id => kanji[id].slug);
 	console.log(learnedKanji);
 	chrome.storage.local.set({"wkhighlight_learnedKanji": learnedKanji, "wkhighlight_learnedKanji_updated":formatDate(new Date())});
 	return learnedKanji;
@@ -94,7 +94,7 @@ const setupContentScripts = (apiToken, learnedKanjiSource, allkanji) => {
 		let allLearnableKanji = result["wkhighlight_allLearnableKanji"];
 		const kanjiList = [];
 		if (!allLearnableKanji) {
-			allLearnableKanji = allkanji["wkhighlight_allkanji"];
+			allLearnableKanji = allkanji;
 			for (let kanjiId in allLearnableKanji) {
 				kanjiList.push(allLearnableKanji[kanjiId]["slug"]);
 			}
@@ -217,107 +217,22 @@ chrome.webNavigation.onDOMContentLoaded.addListener(details => {
 										const vocabUpdates = result["wkhighlight_allvocab_updated"] ? result["wkhighlight_allvocab_updated"] : now;
 										modifiedSince(apiToken, kanjiUpdated, "https://api.wanikani.com/v2/subjects?types=kanji")
 											.then(modified => {
-												if (!result['wkhighlight_allkanji'] || modified) {
-													// fetch all kanji
-													console.log("fetch all kanji");
-													fetchAllPages(apiToken, "https://api.wanikani.com/v2/subjects?types=kanji")
-														.then(kanji_data => {
-															const kanji_dict = {};
-															const kanji_assoc = {};
-															kanji_data.map(content => content.data)
-																.flat(1)
-																.forEach(kanji => {
-																	const data = kanji.data;
-																	kanji_dict[kanji.id] = {
-																		"amalgamation_subject_ids" : data.amalgamation_subject_ids,
-																		"characters" : data.characters,
-																		"component_subject_ids" : data.component_subject_ids,
-																		"document_url" : data.document_url,
-																		"level" : data.level,
-																		"meaning_hint" : data.meaning_hint,
-																		"meaning_mnemonic" : data.meaning_mnemonic,
-																		"meanings" : data.meanings.map(data => data.meaning),
-																		"reading_hint" : data.reading_hint,
-																		"reading_mnemonic" : data.reading_mnemonic,
-																		"readings" : data.readings,
-																		"visually_similar_subject_ids" : data.visually_similar_subject_ids,
-																		"slug": data.slug,
-																		"id":kanji.id,
-																		"subject_type":kanji.object
-																	};
-																	kanji_assoc[data.slug] = kanji.id;
-																});
-															
-															setupContentScripts(apiToken, "https://api.wanikani.com/v2/assignments", {"wkhighlight_allkanji":kanji_dict});
-															// saving all kanji
-															chrome.storage.local.set({"wkhighlight_allkanji": kanji_dict, "wkhighlight_kanji_assoc": kanji_assoc, "wkhighlight_allkanji_updated": now});
-														})
-														.catch(errorHandling);
-												}
-												else {
-													setupContentScripts(apiToken, "https://api.wanikani.com/v2/assignments", result);
-												}
+												if (!result['wkhighlight_allkanji'] || modified)
+													setupKanji(apiToken, "https://api.wanikani.com/v2/subjects?types=kanji", kanji_dict => setupContentScripts(apiToken, "https://api.wanikani.com/v2/assignments", kanji_dict));
+												else
+													setupContentScripts(apiToken, "https://api.wanikani.com/v2/assignments", result["wkhighlight_allkanji"]);
 											});
 										
 										modifiedSince(apiToken, radicalsUpdated, "https://api.wanikani.com/v2/subjects?types=radical")
 											.then(modified => {
-												if (!result['wkhighlight_allradicals'] || modified) {
-													// fetch all radicals
-													fetchAllPages(apiToken, "https://api.wanikani.com/v2/subjects?types=radical")
-														.then(radical_data => {
-															const radical_dict = {};
-															radical_data.map(content => content.data)
-																.flat(1)
-																.forEach(radical => {
-																	const data = radical.data;
-																	radical_dict[radical.id] = {
-																		"characters" : data.characters,
-																		"character_images" : data.character_images,
-																		"document_url" : data.document_url,
-																		"level" : data.level,
-																		"id":radical.id,
-																		"meanings": data.meanings.map(data => data.meaning),
-																		"subject_type":radical.object
-																	};
-																});
-	
-															// saving all radical
-															chrome.storage.local.set({"wkhighlight_allradicals": radical_dict, "wkhighlight_allradicals_updated": now});
-														})
-														.catch(errorHandling);
-												}
+												if (!result['wkhighlight_allradicals'] || modified)
+													setupRadicals(apiToken, "https://api.wanikani.com/v2/subjects?types=radical");
 											});
 	
 										modifiedSince(apiToken, vocabUpdates, "https://api.wanikani.com/v2/subjects?types=vocabulary")
 											.then(modified => {
-												if (!result['wkhighlight_allvocab'] || modified) {
-													fetchAllPages(apiToken, "https://api.wanikani.com/v2/subjects?types=vocabulary")
-														.then(vocab => {
-															const vocab_dict = {};
-															vocab.map(content => content.data)
-																.flat(1)
-																.forEach(vocab => {
-																	const data = vocab.data;
-																	vocab_dict[vocab.id] = {
-																		"characters" : data.characters,
-																		"component_subject_ids" : data.component_subject_ids, 
-																		"context_sentences" : data.context_sentences,
-																		"document_url" : data.document_url,
-																		"level" : data.level,
-																		"meaning_mnemonic" : data.meaning_mnemonic,
-																		"meanings" : data.meanings.map(data => data.meaning),
-																		"parts_of_speech" : data.parts_of_speech,
-																		"reading_mnemonic" : data.reading_mnemonic,
-																		"readings" : data.readings.map(data => data.reading),
-																		"pronunciation_audios" : data.pronunciation_audios,
-																		"id":vocab.id,
-																		"subject_type":vocab.object
-																	};
-																});
-															chrome.storage.local.set({'wkhighlight_allvocab':vocab_dict, "wkhighlight_allvocab_updated": now});
-														})
-														.catch(errorHandling);
-												}
+												if (!result['wkhighlight_allvocab'] || modified)
+													setupVocab(apiToken, "https://api.wanikani.com/v2/subjects?types=vocabulary");
 											});
 									});
 	
@@ -461,39 +376,40 @@ chrome.storage.local.get(["wkhighlight_settings"], result => {
 
 chrome.alarms.onAlarm.addListener(alarm => {
 	if (alarm.name === "next-reviews") {
-		chrome.storage.local.get(["wkhighlight_reviews", "wkhighlight_next-reviews-bundle", "wkhighlight_apiKey"], result => {
-			const reviews = result["wkhighlight_reviews"];
-			const reviews_bundle = result["wkhighlight_next-reviews-bundle"];
-			if (reviews_bundle && reviews && reviews["count"]) {
-				// notify user
-				chrome.notifications.create({
-					type: "basic",
-					title: `WaniKani: ${reviews_bundle.length} new Reviews`,
-					message: `You have now ${reviews_bundle.length} more Reviews, a total of ${reviews["count"]+reviews_bundle.length} Reviews.`,
-					iconUrl: "../logo/logo.png"
-				});
-			}
+		chrome.storage.local.get(["wkhighlight_next-reviews-bundle", "wkhighlight_apiKey"], result => {
+			setupAvailableAssignments(result["wkhighlight_apiKey"], (reviews, lessons) => {
+				const reviews_bundle = result["wkhighlight_next-reviews-bundle"];
+				if (reviews_bundle && reviews && reviews["count"]) {
+					// notify user
+					chrome.notifications.create({
+						type: "basic",
+						title: `WaniKani: ${reviews_bundle.length} new Reviews`,
+						message: `You have now ${reviews_bundle.length} more Reviews, a total of ${reviews["count"]+reviews_bundle.length} Reviews.`,
+						iconUrl: "../logo/logo.png"
+					});
+				}
 
-			// setup new alarm
-			updateAvailableAssignments(result["wkhighlight_apiKey"], setupReviewsAlarm);
+				// setup new alarm
+				setupAvailableAssignments(result["wkhighlight_apiKey"], setupReviewsAlarm);
+			});
 		});	
 	}
 
 	if (alarm.name === "practice") {
-		chrome.storage.local.get(["wkhighlight_lessons", "wkhighlight_reviews", "wkhighlight_practice_timestamp"], result => {
-			const lessons = result["wkhighlight_lessons"]?.count;
-			const reviews = result["wkhighlight_reviews"]?.count;
-			if (lessons!==undefined && reviews!==undefined) {
-				chrome.notifications.create({
-					type: "basic",
-					title: "Time to practice your Japanese!",
-					message: `Lessons: ${lessons}  /  Reviews: ${reviews}`,
-					iconUrl: "../logo/logo.png"
-				});
-			}
-
-			// setup new alarm
-			setupPracticeAlarm(result["wkhighlight_practice_timestamp"]);
+		chrome.storage.local.get(["wkhighlight_apiKey", "wkhighlight_practice_timestamp"], result => {
+			setupAvailableAssignments(result["wkhighlight_apiKey"], (reviews, lessons) => {
+				if (lessons!==undefined && reviews!==undefined) {
+					chrome.notifications.create({
+						type: "basic",
+						title: "Time to practice your Japanese!",
+						message: `Lessons: ${lessons["count"]}  /  Reviews: ${reviews["count"]}`,
+						iconUrl: "../logo/logo.png"
+					});
+				}
+	
+				// setup new alarm
+				setupPracticeAlarm(result["wkhighlight_practice_timestamp"]);
+			});
 		});
 	}
 });